+++ src/main/java/lu/crx/financing/entities/Invoice.java
@@ -1,6 +1,7 @@
 package lu.crx.financing.entities;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import jakarta.persistence.Basic;
 import jakarta.persistence.Entity;
@@ -60,12 +61,12 @@ public class Invoice implements Serializable {
      * The early payment amount of the invoice. Should be not null if invoice has been financed.
      */
     @Basic
-    private Long earlyPaymentAmountInCents;
+    private BigDecimal earlyPaymentAmountInCents;
 
     /**
      * The early discounted amount of the invoice. Should be not null if invoice has been financed.
      *  The sum earlyPaymentAmountInCents + discountedAmountInCents should be equal to valueInCents.
      */
     @Basic
-    private Long discountedAmountInCents;
+    private BigDecimal discountedAmountInCents;
 }
