+++ src/main/java/lu/crx/financing/entities/InvoiceFinancingDetails.java
@@ -1,6 +1,7 @@
 package lu.crx.financing.entities;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import jakarta.persistence.Basic;
@@ -61,17 +62,17 @@ public class InvoiceFinancingDetails implements Serializable {
      * The applied financing rate in bps.
      */
     @Basic(optional = false)
-    private int financingRateInBps;
+    private BigDecimal financingRateInBps;
 
     /**
      * Early payment amount in cents.
      */
     @Basic(optional = false)
-    private long earlyPaymentAmountInCents;
+    private BigDecimal earlyPaymentAmountInCents;
 
     /**
      * Discounted amount in cents.
      */
     @Basic(optional = false)
-    private long discountedAmountInCents;
+    private BigDecimal discountedAmountInCents;
 }
